+++ python/tvm/relay/frontend/tensorflow.py
@@ -3870,21 +3870,13 @@ def from_tensorflow(graph, layout="NHWC", shape=None, outputs=None):
     g = GraphProto()
     
 
-    def detect_tf2_control_flow(_graph):
-        tensorlist_ops = ["TensorListFromTensor", "TensorListGetItem", "TensorListReserve", "TensorListSetItem", "TensorListStack"]
-        control_flow_ops = ["If", "StatelessIf", "While", "StatelessWhile"]
-        for node in _graph.node:
-            # TODO: how about shape?
-            if node.op in tensorlist_ops+control_flow_ops:
-                return True
-        return False
-
-    tf2_needed = detect_tf2_control_flow(graph)
+    from tvm.relay.frontend.tensorflow2 import detect_tf2_control_flow as _detect_tf2_control_flow
+    tf2_needed = _detect_tf2_control_flow(graph)
 
     if not tf2_needed:
         mod, params = g.from_tensorflow(graph, layout, shape, outputs)
     else:
-        from tvm.relay.frontend.tensorflow2 import from_tensorflow as from_tensorflow2
-        mod, params = from_tensorflow2(graph, layout, shape, outputs)
+        from tvm.relay.frontend.tensorflow2 import from_tensorflow as _from_tensorflow2
+        mod, params = _from_tensorflow2(graph, layout, shape, outputs)
 
     return mod, params
